Refactor FormResponse and HTTPResponse
