+++ www/core/Protocols/Connect/ConnectModule.php
@@ -26,13 +26,13 @@ use Fernet\Fernet;
 use Psr\Log\LogLevel;
 use SimpleID\Auth\AuthManager;
 use SimpleID\ModuleManager;
+use SimpleID\Protocols\HTTPResponse;
 use SimpleID\Protocols\ProtocolResult;
 use SimpleID\Protocols\OAuth\OAuthModule;
 use SimpleID\Protocols\OAuth\OAuthProtectedResource;
 use SimpleID\Protocols\OAuth\OAuthDynamicClient;
 use SimpleID\Protocols\OAuth\Response;
 use SimpleID\Store\StoreManager;
-use SimpleID\Util\HTTPResponse;
 use SimpleJWT\Util\Helper;
 use SimpleJWT\JWT;
 use SimpleJWT\InvalidTokenException;
