+++ www/core/Protocols/HTTPResponse.php
@@ -20,7 +20,7 @@
  * 
  */
 
-namespace SimpleID\Util;
+namespace SimpleID\Protocols;
 
 /**
  * Class representing a response from a HTTP request made using the
