+++ www/core/Protocols/OAuth/OAuthDynamicClient.php
@@ -22,8 +22,8 @@
 
 namespace SimpleID\Protocols\OAuth;
 
+use SimpleID\Protocols\HTTPResponse;
 use SimpleID\Crypt\Random;
-use SimpleID\Util\HTTPResponse;
 
 /**
  * A class representing an OAuth dynamic client.
